requirements.txt
flask
spacy
transformers
torch
pdfminer.six
python-docx
